refactor(linear): dedup two-view solvers onto vision-geometry (C1-FOLLOWUP, 1b)
Delete linear's duplicated epipolar / homography / triangulation / camera_matrix
modules and route all callers to vision-geometry, the canonical (superior)
implementations: Hartley-normalized 8-point, AtA-null-space DLT (no SVD hang),
plus collinearity / coplanarity / rank guards. linear now depends on
vision-geometry and bridges errors via Error: From<GeometryError>.

- Zhang init (iterative_intrinsics) + Scheimpflug init + pipeline analysis now
  call vision_geometry::homography::dlt_homography; gated on the real-data golden
  tests (planar_intrinsics_real_data, stereo_linear) — green within tolerance.
- Facade: two-view solvers move from `linear` to a new vision_calibration::geometry
  module (clean pre-1.0 path break).
- stereo_linear: fundamental_8point re-pinned 0.07 -> 0.09 (geometry normalizes;
  the old linear solver did not — ~0.077 vs the analytic golden F, a noise-level
  shift). Golden F kept as ground truth.
- Ported the dense-225 homography perf-regression guard into vision-geometry.

Net -1811 LoC of duplicated solver code. vision-mvg unaffected (never used linear).
