+++ crates/vision-calibration-linear/Cargo.toml
@@ -16,6 +16,7 @@ workspace = true
 
 [dependencies]
 vision-calibration-core = { workspace = true }
+vision-geometry = { workspace = true }
 nalgebra = { workspace = true }
 log = { workspace = true }
 serde = { workspace = true }
