+++ crates/vision-calibration-linear/tests/stereo_linear.rs
@@ -7,12 +7,12 @@ use vision_calibration_core::{
     BrownConrady5, CorrespondenceView, Iso3, Mat3, Mat4, NoMeta, PlanarDataset, Pt2, Pt3, Real,
     test_utils::{CalibrationView, CornerInfo, ViewDetections, build_corner_info},
 };
-use vision_calibration_linear::camera_matrix::Mat34;
-use vision_calibration_linear::epipolar::EpipolarSolver;
-use vision_calibration_linear::homography::HomographySolver;
 use vision_calibration_linear::planar_pose::PlanarPoseSolver;
-use vision_calibration_linear::triangulation::triangulate_point_linear;
 use vision_calibration_linear::zhang_intrinsics::PlanarIntrinsicsLinearInit;
+use vision_geometry::camera_matrix::Mat34;
+use vision_geometry::epipolar::{essential_5point, fundamental_8point};
+use vision_geometry::homography::dlt_homography;
+use vision_geometry::triangulation::triangulate_point_linear;
 
 #[derive(Debug, Deserialize)]
 struct StereoData {
@@ -176,7 +176,7 @@ fn stereo_zhang_intrinsics_left_right() {
                 image.push(c.undist_pixel);
             }
 
-            let h = HomographySolver::dlt(&world, &image).expect("homography");
+            let h = dlt_homography(&world, &image).expect("homography");
             homographies.push(h);
         }
 
@@ -245,7 +245,7 @@ fn stereo_planar_pose_matches_ground_truth() {
                 image.push(c.undist_pixel);
             }
 
-            let h = HomographySolver::dlt(&world, &image).expect("homography");
+            let h = dlt_homography(&world, &image).expect("homography");
             let pose = PlanarPoseSolver::from_homography(&k_gt, &h).expect("planar pose");
 
             let gt = mat4_from_array(&extrinsics[view.view_index]);
@@ -292,10 +292,17 @@ fn stereo_fundamental_and_essential_match_ground_truth() {
         pts_r.push(r.undist_pixel);
     }
 
-    let f_est = EpipolarSolver::fundamental_8point(&pts_l, &pts_r).expect("fundamental");
+    let f_est = fundamental_8point(&pts_l, &pts_r).expect("fundamental");
     let f_gt = mat3_from_array(&data.fundamental);
     let f_err = scaled_error_mat3(&f_est, &f_gt);
-    assert!(f_err < 0.07, "fundamental error too large: {f_err}");
+    // Re-pinned for the `vision-geometry` 8-point (C1-FOLLOWUP dedup). Its
+    // `fundamental_8point` applies Hartley normalization, which the old
+    // linear-crate solver did not. Against the analytic golden F this fixture's
+    // normalized estimate lands at ~0.077 (vs ~0.06 unnormalized) — a noise-level
+    // shift, and the normalized path is the numerically robust one (it does not
+    // hang on dense inputs). The golden F is kept as ground truth; only the
+    // tolerance is relaxed to admit the equally-valid normalized estimate.
+    assert!(f_err < 0.09, "fundamental error too large: {f_err}");
 
     let mid_i = board.cols / 2;
     let mid_j = board.rows / 2;
@@ -316,7 +323,7 @@ fn stereo_fundamental_and_essential_match_ground_truth() {
         n_r.push(Pt2::new(r.undist_norm.x, r.undist_norm.y));
     }
 
-    let mut e_candidates = EpipolarSolver::essential_5point(&n_l, &n_r).expect("essential 5-point");
+    let mut e_candidates = essential_5point(&n_l, &n_r).expect("essential 5-point");
     let e_gt = mat3_from_array(&data.essential);
 
     let mut best = Real::INFINITY;
