+++ crates/vision-calibration-optim/Cargo.toml
@@ -31,6 +31,7 @@ schemars = { workspace = true, optional = true }
 [dev-dependencies]
 vision-calibration-core = { workspace = true, features = ["test-utils"] }
 vision-calibration-linear = { workspace = true }
+vision-geometry = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 criterion = { workspace = true }
