+++ crates/vision-calibration-pipeline/Cargo.toml
@@ -24,6 +24,7 @@ schemars = [
 [dependencies]
 vision-calibration-core = { workspace = true }
 vision-calibration-linear = { workspace = true }
+vision-geometry = { workspace = true }
 vision-calibration-optim = { workspace = true }
 vision-calibration-dataset = { workspace = true }
 vision-calibration-detect = { workspace = true }
